rename sbs and make module
