+++ omc3/segment_by_segment/phase_writer.py
